backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
mcp
httpx